Update zshrc - ensure `brew update` reports all via API
A while back, homebrew was updated to not report all formulae changes
via API. A change was made that if the
`HOMEBREW_UPDATE_REPORT_ALL_FORMULAE` was enabled but
`HOMEBREW_INSTALL_FROM_API` was not, it would result in an error.

Since I like seeing all the new/updated formulae/casks as it is my way
of discovering new software/tools, I made the requisite changes to
enable this.

Respective homebrew PRs:

- [Note that brew update will only report all without API](https://github.com/Homebrew/brew/pull/14866)
- [Document HOMEBREW_INSTALL_FROM_API](https://github.com/Homebrew/brew/pull/12305)