[ci] Switch to timewarrior to speed up Github Action
Github Action run time increased from ~3 minutes to ~13 minutes due to
changing from previous `neovim` role to `youtube-dl` role. This was due
to `neovim` being consolidated into the `common-cli` role which includes
tons of packages which are installed for multiple OSs.

For our test purposes, we need something light to install and verify so
switching to `timewarrior` as that role only includes installing the
package via the respective package manager.